refactor: ResourceSchema 타입 파일 위치 수정
