+++ src/features/resource-list/templates/ResourceUrlTemplate.tsx
@@ -9,8 +9,8 @@ import { toast } from "react-toastify"
 import { styled } from "../../../stitches.config"
 import { Input } from "../../common/components/Input"
 import { VStack } from "../../common/components/Stack"
+import { ResourceSchema } from "../../resource/type"
 import { ResourceAddButton } from "../components/ResourceAddButton"
-import { ResourceSchema } from "../type"
 import { controlValidation } from "../utils/controlValidation"
 import { convertUrl } from "../utils/convertUrl"
 import { validateUrl } from "../utils/validateUrl"
