+++ src/features/resource/type.ts
